updated the variable for movie profits from int32 to float and modified web.config to use a supplimentary config file for connection strings that the gitignore file will not add to the remote repo

+++ herman v2/Models/tmdbVideoDetails.cs
@@ -98,7 +98,7 @@ namespace herman_v2.Models
         public List<ProductionCompany> production_companies { get; set; }
         public List<ProductionCountry> production_countries { get; set; }
         public string release_date { get; set; }
-        public int revenue { get; set; }
+        public long revenue { get; set; }
         public int runtime { get; set; }
         public List<SpokenLanguage> spoken_languages { get; set; }
         public string status { get; set; }
